Manage case status with dedicated controller
