+++ app/views/investigations/status/close.html.erb
@@ -1,4 +1,4 @@
-<%= page_title "Re-open case", errors: @change_case_status_form.errors.any? %>
+<%= page_title "Close case", errors: @change_case_status_form.errors.any? %>
 <% content_for :back_link do %>
   <%= govukBackLink(
     text: "Back to #{@investigation.pretty_description.downcase}",
@@ -8,7 +8,7 @@
 
 <%= render "investigations/pages_top", investigation: @investigation %>
 
-<%= form_with(model: @change_case_status_form, url: reopen_investigation_path(@investigation), local: true, method: :patch) do |form| %>
+<%= form_with(model: @change_case_status_form, url: close_investigation_status_path(@investigation), local: true, method: :patch) do |form| %>
   <div class="govuk-grid-row">
     <div class="govuk-grid-column-two-thirds-from-desktop">
       <%= render "form_components/govuk_error_summary", form: form %>
@@ -17,10 +17,10 @@
                  key: :rationale,
                  form: form,
                  attributes: { maxlength: 10_000 },
-                 label: { text: "Why are you re-opening the case?", classes: "govuk-label--l", isPageHeading: true },
-                 hint: { text: "For example, the business responsible needs to take further action." }
+                 label: { text: "Why are you closing the case?", classes: "govuk-label--l", isPageHeading: true },
+                 hint: { text: "For example, the product is low risk, so no further activity is expected." }
       %>
-      <%= form.submit "Re-open case", class: "govuk-button" %>
+      <%= form.submit "Close case", class: "govuk-button" %>
 
       <p><%= link_to "Cancel", investigation_path(@investigation), class: "govuk-link--no-visited-state" %></p>
     </div>
