+++ spec/requests/investigations/updating_the_status_of_a_case_spec.rb
@@ -23,7 +23,7 @@ RSpec.describe "Updating the status of a case", :with_stubbed_elasticsearch, :wi
     before do
       sign_in user_from_owner_team
       User.current = user_from_owner_team
-      patch close_investigation_path(investigation),
+      patch close_investigation_status_path(investigation),
             params: {
               change_case_status_form: { rationale: "Test" }
             }
@@ -39,7 +39,7 @@ RSpec.describe "Updating the status of a case", :with_stubbed_elasticsearch, :wi
 
     context "when case is re-opened" do
       before do
-        patch reopen_investigation_path(investigation),
+        patch reopen_investigation_status_path(investigation),
               params: {
                 change_case_status_form: { rationale: "Test" }
               }
@@ -57,7 +57,7 @@ RSpec.describe "Updating the status of a case", :with_stubbed_elasticsearch, :wi
     before do
       sign_in user_from_collaborator_team
       User.current = user_from_collaborator_team
-      patch close_investigation_path(investigation),
+      patch close_investigation_status_path(investigation),
             params: {
               change_case_status_form: { rationale: "Test" }
             }
@@ -75,7 +75,7 @@ RSpec.describe "Updating the status of a case", :with_stubbed_elasticsearch, :wi
     before do
       sign_in other_user
       User.current = other_user
-      patch close_investigation_path(investigation),
+      patch close_investigation_status_path(investigation),
             params: {
               change_case_status_form: { rationale: "Test" }
             }
