+++ spec/support/features/page_expectations.rb
@@ -182,12 +182,12 @@ module PageExpectations
   end
 
   def expect_to_be_on_close_case_page(case_id:)
-    expect(page).to have_current_path("/cases/#{case_id}/close")
+    expect(page).to have_current_path("/cases/#{case_id}/status/close")
     expect(page).to have_h1("Why are you closing the case?")
   end
 
   def expect_to_be_on_reopen_case_page(case_id:)
-    expect(page).to have_current_path("/cases/#{case_id}/reopen")
+    expect(page).to have_current_path("/cases/#{case_id}/status/reopen")
     expect(page).to have_h1("Why are you re-opening the case?")
   end
